requirements.txt
Flask~=1.1.2
Flask-MySQL
Flask-Cors
Pandas~=1.1.1
mysql-connector-python~=8.0.21
google-cloud-bigquery